+++ typescript/sdk/src/ism/read.ts
@@ -13,13 +13,13 @@ import {
   Address,
   WithAddress,
   assert,
+  concurrentMap,
   ethersBigNumberSerializer,
   rootLogger,
 } from '@hyperlane-xyz/utils';
 
-import { Chains } from '../consts/chains.js';
-import { concurrentMap } from '../index.js';
 import { MultiProvider } from '../providers/MultiProvider.js';
+import { ChainName } from '../types.js';
 
 import {
   AggregationIsmConfig,
@@ -53,7 +53,7 @@ export class EvmIsmReader implements IsmReader {
 
   constructor(
     protected readonly multiProvider: MultiProvider,
-    chain: Chains,
+    chain: ChainName,
     protected readonly concurrency: number = 20,
   ) {
     this.provider = this.multiProvider.getProvider(chain);
